scheduler: move Describe map-clone test into scheduler_test.go
Address review: consolidate the test into the package's existing
scheduler_test.go rather than a separate file.
